requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
bitarray==3.0.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
ckzg==2.0.1
click==8.1.7
colorama==0.4.6
cytoolz==1.0.0
Django==5.1.3
eth-account==0.13.4
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.6.0
eth-rlp==2.1.0
eth-typing==5.0.1
eth-utils==5.1.0
eth_abi==5.1.0
Flask==3.1.0
frozenlist==1.5.0
hexbytes==1.2.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
multidict==6.1.0
packaging==23.2
parsimonious==0.10.0
propcache==0.2.1
py-solc-x==2.0.3
pycryptodome==3.21.0
pydantic==2.10.2
pydantic_core==2.27.1
pyunormalize==16.0.0
pywin32==308
regex==2024.11.6
requests==2.32.3
rlp==4.0.1
sqlparse==0.5.2
toolz==1.0.0
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
web3==7.6.0
websockets==13.1
Werkzeug==3.1.3
yarl==1.18.3
